requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pytz==2018.9
sqlparse==0.2.4
